Remove the use of React.memo for now
This leads to React reporting 'Invalid Hook Call', likely due to the use of the `useStyles`
hook inside the memo. The proposed fix on https://fb.me/react-invalid-hook-call is to switch
the order of React.memo and useStyles, for now just remove this in preparation of some more
feature work that should go in first.