Update the eslint configuration to allow editors to understand it
Explicitly listing the .storybook/stories path in the package.json's `test:lint` invocation
works for ensuring that things get linted as part of the CI process, but VSCode (or any other
editor, for that matter) isn't aware of that and won't flag issues in these files. This leads
to multiple "fix warnings" commits and releases that are not working.

Instead this commit changes the .eslintignore file to explicitly un-ignore the .storybook directory
as advised by https://github.com/eslint/eslint/issues/12348#issuecomment-536946429.

Note that this commit also fixes a bunch of now-reported warnings in other files that previously
got excluded.