Remove the columns from the component state
The Board uses the columns provided in the props, and will communicate changes to the columns
(or the cards) through the callbacks -- in other words, this is a "controlled component".
As such it doesn't need to maintain any specific state, and rather simply relies on the caller
to provide the callbacks and apply the changes in such a way that the `columns` prop changes.

The API of the component is not changed by this commit, as we previously had all callbacks as optional
and they are now still optional, BUT: If there are no callbacks provided (either the `onChange`
or the more granular `on{Card,Column}{Added,Moved}` callbacks) then the component will no longer
show the results of the user interaction.