Show how a completely "view-only" board would look like
The "read-only" story doesn't provide the `createColumn`/`createCard` helpers, so we do
not show the controls for adding columns and cards -- but it does allow to move columns
and cards around because there is handlers for the changes.

For a view-only board we need to also not provide the change handlers, which works but
leaves the drag/drop functionality enabled to produce an odd experience for the user.
This commit therefore checks whether it makes sense to enable dragging of the columns
and cards when there is handlers for that functionality.